Make Search Even Better  :open_mouth:
How do you make search better you ask?

Well do I have an answer for you!

The first iteration of search used Substring Logic which was good at partial matches but was unforgiving in terms of spelling errors

The second iteration changed that with Soundex Logic. It was now possible to make spelling mistakes and still get the correct product as long as the search term sounded like the product name. But this implementation did not consider substrings

This third iteration of Search integrates the Substring Logic with the Soundex Logic making search so much better.